+++ src/config/index.ts
@@ -1,4 +1,3 @@
-
 // monaco
 export const monacoLoader = 'https://unpkg.com/monaco-editor@0.20.0/min/vs/loader.js'
 export const monaco = 'https://unpkg.com/monaco-editor@0.20.0/min/vs'